Added play command to new restructing
